src/web_operator/agents/agent_state.py
from typing import TypedDict, Annotated, List
import operator
from langchain_core.messages import AnyMessage


class AgentState(TypedDict):
    message: Annotated[List[str], operator.add]
    supervisor_msg:Annotated[List[str], operator.add]
    sender: Annotated[List[str], operator.add]